+++ src/routes/auth/login/+page.svelte
@@ -3,7 +3,7 @@
 	import {page} from '$app/state'
 	import * as m from '$lib/paraglide/messages'
 	import AuthLogin from '$lib/components/auth-login.svelte'
-	import IconR4 from '$lib/icon-r4.svelte'
+	import IconR4 from '$lib/components/icon-r4.svelte'
 
 	const redirect = $derived(page.url.searchParams.get('redirect') || '/settings')